split out MetricsDto.hpp
